CareLogG8/requirements.txt
cryptography
streamlit
pandas
google-generativeai
